+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Removed
+
+- `order-coupon` dependency.
+
 ## [0.7.0] - 2019-11-06
 
 ### Changed